Initial commit
I have brought the core functionality of the Label Retriever from the previous Unity project into a stand alone assembly. The intention with this is that this core functionality can be reused in other applications and therefore should not have to be rewritten in a thousand different locations.

This project is also my first that will automatically generate a NuGet package on build.